# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Agenda de Contacts",
        "description": "Documentação da API de Agenda de Contacts",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/exports": {
            "get": {
                "tags": [
                    "Exportações"
                ],
                "summary": "Lista todas as exportações do usuário autenticado",
                "operationId": "d25d334b0d44e07cb893c96a709ba949",
                "responses": {
                    "200": {
                        "description": "Lista de exportações retornada com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exportações"
                ],
                "summary": "Solicita uma nova exportação de Contacts",
                "operationId": "d729cf8e078f61999ccf50d4359d8de3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "formato"
                                ],
                                "properties": {
                                    "formato": {
                                        "description": "Formato do arquivo para exportação",
                                        "type": "string",
                                        "example": "csv",
                                        "enum": [
                                            "csv",
                                            "xlsx",
                                            "pdf"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exportação solicitada com sucesso"
                    },
                    "422": {
                        "description": "Formato inválido"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exports/{id}": {
            "get": {
                "tags": [
                    "Exportações"
                ],
                "summary": "Consulta o status de uma exportação",
                "operationId": "f1f9f4d065d36f6e95adecce1f82fa50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da exportação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status da exportação retornado com sucesso"
                    },
                    "404": {
                        "description": "Exportação não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/exports/{id}/download": {
            "get": {
                "tags": [
                    "Exportações"
                ],
                "summary": "Realiza o download da exportação concluída",
                "operationId": "9e1f5a8ae212030379780e7af1636c9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da exportação",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Arquivo baixado com sucesso"
                    },
                    "400": {
                        "description": "A exportação ainda não foi concluída"
                    },
                    "404": {
                        "description": "Exportação não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Autenticação via Laravel Sanctum. O frontend usa sessão de SPA (cookie + CSRF); integrações externas podem usar um token pessoal no formato: Bearer <token>.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}